+++ packages/eyes-storybook/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add `visualGridOptions`, `useDom`, and `enablePatterns` config params
 
 ## 3.23.0 - 2021/8/17